feat: allow no auth
License: MIT
