+++ utils.go
@@ -1,0 +1,16 @@
+package webdav
+
+import (
+	"strings"
+
+	"golang.org/x/crypto/bcrypt"
+)
+
+func checkPassword(saved, input string) bool {
+	if strings.HasPrefix(saved, "{bcrypt}") {
+		savedPassword := strings.TrimPrefix(saved, "{bcrypt}")
+		return bcrypt.CompareHashAndPassword([]byte(savedPassword), []byte(input)) == nil
+	}
+
+	return saved == input
+}
